Only append compilation timestamp in debug mode
By vendoring a timestamp into the liveness agent every time we compile
it we're changing the contents unnecessarily. This has the effect of
causing us to rewrite the file and restart the liveness agent service on
all machines in the fleet for every recipe release. This change allows
us to compile the timestamp during development or debug but ommits it
when building a release artifact.