Allow setting config interval and reading from env
Some service providers won't pass `INTERVAL=2` in the tests from the
chef-client environment to the liveness agent's environment.

This leaves the agent on an 1800 second interval causing one of the
tests to fail (test-002 after setup-002 resets the pings).

This change allows interval to be read from the config.json, as well as
will write an interval to the config.json if it is seen in the
environment.

Signed-off-by: Bryan McLellan <btm@loftninjas.org>