Chef-15 fixes for `node[:chef_guid]`
The public API for this now has nothing to do with the DataCollector.

In Chef-15 the `node[:chef_guid]` is created for every single Chef-15
node, doesn't matter if the node is using the data collector or not.

The DataCollector then consumes that.

All the internals of the DataCollector have changed and reading the
guid out of the DataCollector is no longer a public UI of the DC.

The crazysauce with the multiple metadata files is also gone, along with the
data_collector_metadata.json file itself (which the prior PR #67 here
independently discovered).

Included backcompat here for Chef < 15.0 implementations which still
have the Chef::DataCollector::Messages class.